# packages/api-contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/providers/profile": {
      "post": {
        "operationId": "ProvidersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me": {
      "get": {
        "operationId": "ProvidersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      },
      "patch": {
        "operationId": "ProvidersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
...
    "/api/v1/providers/me/skills": {
      "get": {
        "operationId": "ProvidersController_skills",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      },
      "post": {
        "operationId": "ProvidersController_addSkill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProviderSkillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/services": {
      "get": {
        "operationId": "ProvidersController_services",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      },
      "post": {
        "operationId": "ProvidersController_addService",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProviderServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/services/{serviceId}": {
      "delete": {
        "operationId": "ProvidersController_removeService",
        "parameters": [
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/service-areas": {
      "get": {
        "operationId": "ProvidersController_serviceAreas",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      },
      "post": {
        "operationId": "ProvidersController_addServiceArea",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddServiceAreaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/service-areas/{areaId}": {
      "delete": {
        "operationId": "ProvidersController_removeServiceArea",
        "parameters": [
          {
            "name": "areaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/availability": {
      "get": {
        "operationId": "ProvidersController_listAvailability",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      },
      "post": {
        "operationId": "ProvidersController_addAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/availability/{availabilityId}": {
      "delete": {
        "operationId": "ProvidersController_removeAvailability",
        "parameters": [
          {
            "name": "availabilityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/me/skills/{skillId}": {
      "delete": {
        "operationId": "ProvidersController_removeSkill",
        "parameters": [
          {
            "name": "skillId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
...
    "/api/v1/providers/{userId}/approve": {
      "post": {
        "operationId": "ProvidersController_approve",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/{userId}/reject": {
      "post": {
        "operationId": "ProvidersController_reject",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/{userId}/suspend": {
      "post": {
        "operationId": "ProvidersController_suspend",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
    "/api/v1/providers/{userId}/reinstate": {
      "post": {
        "operationId": "ProvidersController_reinstate",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Providers"]
      }
    },
...
    "/api/v1/ops/dead-letter/{id}/resolve": {
      "post": {
        "operationId": "OpsController_resolveDeadLetter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ops"]
      }
    },
...
    "/api/v1/ops/notification-jobs/{id}/retry": {
      "post": {
        "operationId": "OpsController_retryNotificationJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ops"]
      }
    }
  },
  "info": {
    "title": "Emek Core API",
    "description": "İki taraflı dijital hizmet pazaryeri çekirdek API. Hata kodları: docs/api/error-codes.md",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Firebase ID token (ADR-0016)"
      }
    },
    "schemas": {
...
      "CreateProviderProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProviderProfileDto": {
        "type": "object",
        "properties": {}
      },
      "AddProviderSkillDto": {
        "type": "object",
        "properties": {}
      },
      "AddProviderServiceDto": {
        "type": "object",
        "properties": {}
      },
      "AddServiceAreaDto": {
        "type": "object",
        "properties": {}
      },
      "AddAvailabilityDto": {
        "type": "object",
        "properties": {}
      },
      "RejectProviderDto": {
        "type": "object",
        "properties": {}
      },
      "SuspendProviderDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}